+++ scenarios/NGSIM/peachtree/scenario.py
@@ -4,9 +4,17 @@ from smarts.sstudio import types as t
 from smarts.sstudio import gen_scenario
 
 
+traffic_histories = [
+    t.TrafficHistoryDataset(
+        name=f"peach_{hd}",
+        source_type="NGSIM",
+        input_path=None,  # for example: f"./trajectories-{hd}.txt"
+        speed_limit_mps=28,
+        default_heading=0,
+    )
+    for hd in ["0400pm-0415pm", "1245pm-0100pm"]
+]
+
 gen_scenario(
-    t.Scenario(
-        traffic_histories=["peach_0400-0415.yml", "peach_1245-0100.yml"],
-    ),
-    output_dir=Path(__file__).parent,
+    t.Scenario(traffic_histories=traffic_histories), output_dir=Path(__file__).parent
 )